Moved PluginUserInterfaces up to BasicActivateItems to allow console gui plugin UIs
